GSDumpGUI: went a bit mad on changing type width and signs after finding a signedness bug, and (more importantly for now), handled (kind of) split GIF primitives without crashing (hopefully)
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@4164 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ tools/GSDumpGUI/Library/GSDump/GSData/GIFPacket/GIFReg/GIFRegAD.cs
@@ -6,10 +6,10 @@ namespace GSDumpGUI
 {
     public static class GIFRegAD
     {
-        static public GIFReg Unpack(GIFTag tag, int addr, UInt64 LowData, UInt64 HighData, bool PackedFormat)
+        static public GIFReg Unpack(GIFTag tag, byte addr, UInt64 LowData, UInt64 HighData, bool PackedFormat)
         {
-            int reg = (int)GIFReg.GetBit(HighData, 0, 8);
-            if (reg == (int)GIFRegDescriptor.AD)
+            byte reg = (byte)GIFReg.GetBit(HighData, 0, 8);
+            if (reg == (byte)GIFRegDescriptor.AD)
                 return GIFRegNOP.Unpack(tag, reg, LowData, HighData, PackedFormat);
             return GIFTag.GetUnpack(reg)(tag, reg, LowData, HighData, false);
         }
