+++ tools/GSDumpGUI/Library/GSDump/GSData/GIFPacket/GIFReg/GIFRegNOP.cs
@@ -6,11 +6,11 @@ namespace GSDumpGUI
 {
     public class GIFRegNOP : GIFReg
     {
-        public int addr;
+        public byte addr;
 
-        public GIFRegNOP(int addr, UInt64 LowData, UInt64 HighData, bool PackedFormat) : base(addr, LowData, HighData, PackedFormat) { }
+        public GIFRegNOP(byte addr, UInt64 LowData, UInt64 HighData, bool PackedFormat) : base(addr, LowData, HighData, PackedFormat) { }
 
-        static public GIFReg Unpack(GIFTag tag, int addr, UInt64 LowData, UInt64 HighData, bool PackedFormat)
+        static public GIFReg Unpack(GIFTag tag, byte addr, UInt64 LowData, UInt64 HighData, bool PackedFormat)
         {
             GIFRegNOP nop = new GIFRegNOP(addr, LowData, HighData, PackedFormat);
             nop.Descriptor = GIFRegDescriptor.NOP;
